refactor: move scraper functions to seperate file
